requirements.txt
# Core Dependencies
python-dotenv==1.1.0

# LangChain & LangGraph (Installed versions as of 2025-01-11)
langchain==1.0.3
langchain-core==1.0.2
langchain-openai==1.0.1
langchain-text-splitters==1.0.0

# LangGraph Agent Framework
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-checkpoint-postgres==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9

# Database Drivers
psycopg[binary,pool]==3.2.12
asyncpg==0.30.0

# Supporting Libraries (auto-installed dependencies)
pydantic==2.12.3
tiktoken==0.12.0
openai==2.6.1

# API Server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Production Monitoring (Optional - uncomment to use)
# prometheus-client==0.21.0
# sentry-sdk==2.19.0

# Development Tools (Optional)
# pytest==8.3.4
# pytest-asyncio==0.24.0
# black==25.1.0
# mypy==1.15.0
